Reorganizing tests: separating data simulation from filter tests and categorizing tests by goal or operation.
