+++ build.gradle
@@ -59,6 +59,5 @@ import com.vanniktech.maven.publish.SonatypeHost
 
 mavenPublishing {
     publishToMavenCentral(SonatypeHost.S01)
-
     signAllPublications()
 }
