+++ src/main/java/com/hiperboot/util/DateFormatIdentifier.java
@@ -26,6 +26,10 @@ import lombok.extern.log4j.Log4j2;
 @Log4j2
 public class DateFormatIdentifier {
 
+    private DateFormatIdentifier() {
+
+    }
+
     public static DateFormatType identifyDateFormat(String dateString) {
         try {
             LocalDate.parse(dateString, DateTimeFormatter.ISO_DATE);
