+++ src/test/java/com/hiperboot/data_simulation/repository/hiperboot/book/BookHiperBootRepository.java
@@ -13,13 +13,14 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.hiperboot.db;
+package com.hiperboot.data_simulation.repository.hiperboot.book;
 
-public enum StatusEnum {
-    ON,
-    OFF,
-    HIGH;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
 
-    private StatusEnum() {
-    }
+import com.hiperboot.data_simulation.entity.book.Book;
+import com.hiperboot.db.repository.HiperBootRepository;
+
+@Repository
+public interface BookHiperBootRepository extends HiperBootRepository<Book>, JpaRepository<Book, Long> {
 }
