--- src/test/java/com/hiperboot/db/entity/GranChildTable.java
@@ -1,42 +1,0 @@
-/*
- * Copyright 2002-2024 by Sannon Gualda de Aragão.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.hiperboot.db.entity;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-import jakarta.persistence.Table;
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-@Entity
-@Table(name = "gran_child_table")
-public class GranChildTable {
-
-    @Id
-    @Column(name = "gran_name")
-    private String granName;
-
-    private String something;
-
-    @ManyToOne
-    @JoinColumn(name = "child_table_name", referencedColumnName = "name")
-    private ChildTable childTable;
-}
