+++ src/test/java/com/hiperboot/pagination/BasePageTest.java
@@ -26,8 +26,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import com.hiperboot.BaseTestClass;
 import com.hiperboot.db.domain.BasePage;
 import com.hiperboot.db.domain.Pagination;
-import com.hiperboot.db.entity.book.Book;
-import com.hiperboot.db.repository.hiperboot.book.BookHiperBootRepository;
+import com.hiperboot.data_simulation.entity.book.Book;
+import com.hiperboot.data_simulation.repository.hiperboot.book.BookHiperBootRepository;
 
 class BasePageTest extends BaseTestClass {
